Begin to add Federated tests.
Fix a bug with allocation Collections instead of OrderedCollections for
inbox and outbox test actors.

Add option for webfinger to be supported for the test run.

Fix mapping bugs with the generic web handler and Inbox/Outbox routing
of requests.

Fix putting context keys in for HTTP Signature verification, which now
behaves properly for POSTing to the inbox.